Restructure config into domain-specific folders, add config snapshots to runs
Split eval_config/ into config/ with subfolders: user_ingestion/, document_ingestion/,
document_chunking/, and evaluation/. Each script now imports only from its relevant
config subdirectory. Ontology and custom instructions are split into separate user
vs document files. Each pipeline step snapshots its config into the run output
directory for reproducibility. Removed DOCUMENT_INGEST_LIMIT.
