+++ zep-eval-harness/README.md
@@ -348,7 +348,6 @@ The evaluation script uses `cross_encoder` reranker by default for best accuracy
 - `LLM_RESPONSE_MODEL`: Model for generating responses
 - `LLM_JUDGE_MODEL`: Model for grading answers
 - `LLM_CONTEXTUALIZATION_MODEL`: Model for document chunk contextualization
-- `DOCUMENT_INGEST_LIMIT`: Maximum number of documents to ingest from `data/documents/`. Documents are sorted alphabetically by filename, so the selection is deterministic. Set to `None` to ingest all documents (default). Set to an integer (e.g. `1`) to ingest only the first N documents.
 
 You can experiment with different rerankers by modifying the `reranker` parameter in `perform_graph_search()`:
 - `cross_encoder`: Best accuracy, slower (default)
