+++ zep-eval-harness/config/constants.py
@@ -1,0 +1,6 @@
+# Google Gemini configuration
+GEMINI_BASE_URL = "https://generativelanguage.googleapis.com/v1beta/openai/"
+
+# Polling
+POLL_INTERVAL = 2  # seconds between status checks
+POLL_TIMEOUT = 600  # 10 minutes max wait
