+++ zep-eval-harness/config/document_chunking/constants.py
@@ -1,0 +1,5 @@
+# Document chunking
+CHUNK_SIZE = 500
+
+# Model used for document chunk contextualization
+LLM_CONTEXTUALIZATION_MODEL = "gemini-2.5-flash-lite"
