+++ zep-eval-harness/config/document_ingestion/constants.py
@@ -1,0 +1,2 @@
+# Standalone graph for shared documents (ingested once, accessible to all users)
+DOCUMENTS_GRAPH_ID = "zep_eval_shared_documents"
